[builtins] Avoid using long double in FreeBSD standalone environment
After 05a4212cc76d a number of long double related declarations are
enabled in `int_types.h`, whenever the CPU architecture and platform
support it. However, this does not work with FreeBSD's standalone
environment, which disallows any use of floating point.

In add98b246290 this was made conditional with `CRT_HAS_FLOATING_POINT`,
so use that macro to also determine whether `HAS_80_BIT_LONG_DOUBLE` can
be turned on.